pyproject.toml
[tool.poetry]
name = "diploma-project-baburkin"
version = "0.1.0"
description = ""
authors = ["Artem <a.baburkin@it.adcombo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selene = "2.0.0rc4"
pytest = "7.4.3"
allure-pytest = "^2.13.2"
requests = "^2.31.0"
jsonschema = "^4.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
